Move same_channel tests to own file
